Support AnyCPU
Contains changes to Support dynamic binding on 32 or 64 bit lua implementation depending on the currently running process. Changes only for windows platform. (Implementation only tested on windows, the 32 bit dll name is lua53.dll and for 64 bit processes lua53_64.dll) This behaviour can be changed in NativeMethods.cs (for 32 bit) and NativeMethods64.cs (for 64 bit)